renamed tentacle creator folder and improved tentacle creator experience
