+++ config/cst.py
@@ -179,7 +179,7 @@ TENTACLE_MODULE_TYPE = "type"
 TENTACLE_MODULE_SUBTYPE = "subtype"
 TENTACLE_MODULE_VERSION = "version"
 TENTACLE_MODULE_CONFIG_FILES = "config_files"
-TENTACLE_CREATOR_PATH = "tentacle_creation"
+TENTACLE_CREATOR_PATH = "tentacle_creator"
 TENTACLE_TEMPLATE_DESCRIPTION = "description"
 TENTACLE_TEMPLATE_PATH = "templates"
 TENTACLE_TEMPLATE_PRE_EXT = "_tentacle"
