+++ tools/tentacle_creator/templates/Mode_tentacle.template
