+++ tools/tentacle_creator/templates/RealTime_tentacle.template
