+++ tools/tentacle_creator/templates/Social_tentacle.template
