+++ tools/tentacle_creator/templates/Strategies_tentacle.template
