+++ tools/tentacle_creator/templates/TA_tentacle.template
