+++ tools/tentacle_creator/templates/Util_tentacle.template
