+++ tools/tentacle_creator/templates/description_tentacle.template
